review(#110): nitpick sweep — test placement, exception form, orphan comment
Phase-2 cosmetic fixes:
- Move TestDetectTimeout from test_solve_run.py to test_solver.py (its home
  now that detect_timeout lives in solver.py). test_solve_run keeps the import
  for its slow streaming test.
- solver.py _read_log_text: `except OSError` (FileNotFoundError is a subclass)
  to match solve_run.py's form.
- Drop the now-orphaned detect_timeout justification comment in solve_run.py;
  the solver.detect_timeout(...) call site is self-describing.
