+++ tests/test_seatrades/test_solve_run.py
@@ -42,18 +42,6 @@ class TestPercentFromElapsed:
         assert percent_from_elapsed(90.0, 60.0) == 1.0
 
 
-class TestDetectTimeout:
-    def test_true_when_log_has_time_limit_line(self):
-        """CBC's time-limit line marks the solve as timed out."""
-        log_text = "Cbc0010I After 0 nodes\nResult - Stopped on time limit\nTotal time 60.0"
-        assert detect_timeout(log_text) is True
-
-    def test_false_when_log_lacks_time_limit_line(self):
-        """A normal optimal log is not a timeout."""
-        log_text = "Result - Optimal solution found\nObjective value 42.0"
-        assert detect_timeout(log_text) is False
-
-
 class TestSolveRunIsSideEffectFree:
     def test_init_does_not_start_thread_or_touch_log(
         self, scheduling_problem, isolated_config, sample_assignment_solution
